+++ Engine/src/support_rectangle.cpp
@@ -59,12 +59,12 @@ void Support_rectangle::Draw()
     glEnd();
 }
 
-void Support_rectangle::On_mouse_move(int x, int y)
+void Support_rectangle::On_mouse_move(float x, float y)
 {
     Support_object::On_mouse_move(x, y);
     if (!visible || !selected || is_drag) return;
 
-    vec2 cursor((float)x, (float)y);
+    vec2 cursor(x, y);
 
     if (drag_point) {
         switch (selected_vertex)
@@ -83,7 +83,7 @@ void Support_rectangle::On_mouse_move(int x, int y)
     else selected_vertex = -1;
 }
 
-void Support_rectangle::On_mouse_down(int x, int y, int b)
+void Support_rectangle::On_mouse_down(float x, float y, int b)
 {
     if (!visible || !selected) return;
 
@@ -97,7 +97,7 @@ void Support_rectangle::On_mouse_down(int x, int y, int b)
     }
 }
 
-void Support_rectangle::On_mouse_up(int x, int y, int b)
+void Support_rectangle::On_mouse_up(float x, float y, int b)
 {
     if (drag_point) {
         drag_point = false;
